+++ src/engine/api/json_factory.cpp
@@ -34,10 +34,13 @@ const constexpr char *modifier_names[] = {"uturn",    "sharp right", "right", "s
 // translations of TurnTypes. Not all types are exposed to the outside world.
 // invalid types should never be returned as part of the API
 const constexpr char *turn_type_names[] = {
-    "invalid", "new name",        "continue",        "turn",       "merge",      "ramp",
-    "fork",    "end of road",     "notification",    "roundabout", "roundabout", "rotary",
-    "rotary",  "roundabout_turn", "roundabout_turn", "invalid",    "invalid",    "invalid",
-    "invalid", "invalid",         "invalid",         "invalid",    "invalid",    "invalid"};
+    "invalid",     "new name",     "continue",        "turn",
+    "merge",       "ramp",         "exit_ramp",       "fork",
+    "end of road", "notification", "roundabout",      "roundabout",
+    "rotary",      "rotary",       "roundabout_turn", "roundabout_turn",
+    "invalid",     "invalid",      "invalid",         "invalid",
+    "invalid",     "invalid",      "invalid",         "invalid",
+    "invalid"};
 
 const constexpr char *waypoint_type_names[] = {"invalid", "arrive", "depart"};